refactor(activity): split god-file into focused submodules
`crates/homeboy-core/src/activity.rs` carried 43 top-level items (threshold
30), tripping the structural audit's high-item-count finding (#9794). The file
mixed the serializable data model, multi-source dedup/reconciliation, shared
leaf helpers, three inline source adapters, and report assembly in one place.

Decompose along responsibility boundaries, no behavior change:
- `activity/model.rs` — serializable report types + state predicates.
- `activity/collector.rs` — multi-source dedup, precedence, ref/evidence
  merging, sort, and conflict projection.
- `activity/action_helpers.rs` — shared leaf helpers (timestamp parsing,
  metadata extraction, next-action construction).
- `activity/observation.rs`, `daemon_jobs.rs`, `runner_sessions.rs` — the three
  previously-inline source adapters, now file modules.

The root retains only report assembly and id resolution (15 items); every
resulting file is well under threshold (max 15) and the `activity/` dir holds
7 files (sprawl threshold 50). Public API is preserved via `pub use model::*`,
so `activity::ActivityScope`, `is_active`, `show_activity`, etc. resolve
unchanged for existing consumers.

Closes #9794
