+++ crates/homeboy-core/src/activity.rs
@@ -1,149 +1,37 @@
-use std::collections::{BTreeMap, BTreeSet};
+//! Activity reporting — aggregates in-flight and recent work from observation
+//! runs, agent-task records, daemon jobs, and runner sessions into a single
+//! deduplicated report, and resolves individual items by id.
+//!
+//! The data model lives in [`model`], multi-source dedup/reconciliation in
+//! [`collector`], shared leaf helpers in [`action_helpers`], and each source
+//! adapter in its own submodule (`observation`, `daemon_jobs`,
+//! `runner_sessions`, `agent_task_provider`). This root retains only report
+//! assembly and id resolution (#9794).
 
-use chrono::{DateTime, Utc};
-use serde::{Deserialize, Serialize};
+use std::collections::BTreeSet;
+
 use serde_json::Value;
 
-use crate::api_jobs::{self, ActiveRunnerJobSummary, Job, JobEvent};
-use crate::observation::{ObservationStore, RunListFilter, RunRecord, RunStatus};
-use crate::run_lifecycle_status::RunLifecycleStatus;
-use crate::{paths, Error, Result};
+use crate::{Error, Result};
 
 pub mod agent_task_provider;
 
+mod action_helpers;
+mod collector;
+mod model;
+
+mod daemon_jobs;
+mod observation;
+mod runner_sessions;
+
+pub use model::*;
+
+// Re-exported at module scope so the source-provider submodules can pull the
+// shared helpers and collector through `use super::*` (#9794).
+pub(crate) use action_helpers::{action, metadata_string, ms_to_rfc3339};
+pub(crate) use collector::ActivityCollector;
+
 pub const ACTIVITY_REPORT_SCHEMA: &str = "homeboy/activity-report/v1";
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum ActivityScope {
-    ActiveRecent,
-    All,
-}
-
-pub type ActivityState = RunLifecycleStatus;
-
-pub fn is_active(state: ActivityState) -> bool {
-    matches!(state, ActivityState::Queued | ActivityState::Running)
-}
-
-pub fn is_failure(state: ActivityState) -> bool {
-    !is_active(state) && !matches!(state, ActivityState::Succeeded | ActivityState::Cancelled)
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
-pub struct ActivityNextAction {
-    pub label: String,
-    pub command: String,
-}
-
-#[derive(Debug, Clone, Default, Serialize, Deserialize, PartialEq, Eq)]
-pub struct ActivityRunnerRefs {
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub runner_id: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub job_id: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub transport: Option<String>,
-}
-
-#[derive(Debug, Clone, Default, Serialize, Deserialize, PartialEq, Eq)]
-pub struct ActivityCrossRefs {
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub run_id: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub agent_task_run_id: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub runner_job_id: Option<String>,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
-pub struct ActivityEvidenceRef {
-    pub id: String,
-    pub kind: String,
-    pub uri: String,
-}
-
-/// A store-specific view retained with the canonical activity item so state
-/// reconciliation remains inspectable without returning duplicate work items.
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
-pub struct ActivitySourceProjection {
-    pub source_store: String,
-    pub id: String,
-    pub state: ActivityState,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub updated_at: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub finished_at: Option<String>,
-}
-
-/// A non-authoritative source state retained for operators investigating a
-/// reconciled activity item. The top-level item state remains authoritative.
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
-pub struct ActivityStateConflict {
-    pub source_store: String,
-    pub id: String,
-    pub state: ActivityState,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
-pub struct ActivityItem {
-    pub id: String,
-    pub kind: String,
-    pub source_store: String,
-    pub state: ActivityState,
-    pub created_at: String,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub updated_at: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub finished_at: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub command: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub cwd: Option<String>,
-    #[serde(default)]
-    pub runner: ActivityRunnerRefs,
-    #[serde(default)]
-    pub refs: ActivityCrossRefs,
-    #[serde(default, skip_serializing_if = "Vec::is_empty")]
-    pub artifacts: Vec<ActivityEvidenceRef>,
-    #[serde(default, skip_serializing_if = "Vec::is_empty")]
-    pub evidence: Vec<ActivityEvidenceRef>,
-    #[serde(default, skip_serializing_if = "Vec::is_empty")]
-    pub source_projections: Vec<ActivitySourceProjection>,
-    #[serde(default, skip_serializing_if = "Vec::is_empty")]
-    pub state_conflicts: Vec<ActivityStateConflict>,
-    #[serde(default, skip_serializing_if = "Vec::is_empty")]
-    pub next_actions: Vec<ActivityNextAction>,
-}
-
-#[derive(Debug, Clone, Default, Serialize, Deserialize, PartialEq, Eq)]
-pub struct ActivityCounts {
-    pub total: usize,
-    pub active: usize,
-    pub queued: usize,
-    pub running: usize,
-    pub succeeded: usize,
-    pub partial_failure: usize,
-    pub failed: usize,
-    pub cancelled: usize,
-    pub timed_out: usize,
-    pub stale: usize,
-    pub unknown: usize,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
-pub struct ActivityReport {
-    pub schema: &'static str,
-    pub command: &'static str,
-    pub counts: ActivityCounts,
-    pub items: Vec<ActivityItem>,
-    /// Agent-task record-health summary, carried as JSON so core does not depend
-    /// on the agent-task health type. Supplied by the agent-task activity
-    /// provider (null when the agent-task subsystem is absent).
-    #[serde(default)]
-    pub agent_task_record_health: Value,
-    #[serde(default, skip_serializing_if = "Vec::is_empty")]
-    pub next_actions: Vec<String>,
-}
 
 pub fn activity_report(scope: ActivityScope, limit: usize) -> Result<ActivityReport> {
     let mut collector = ActivityCollector::default();
@@ -269,559 +157,12 @@ fn counts_for_items(items: &[ActivityItem]) -> ActivityCounts {
     counts
 }
 
-#[derive(Default)]
-struct ActivityCollector {
-    items: BTreeMap<String, ActivityItem>,
-}
-
-impl ActivityCollector {
-    fn insert(&mut self, mut item: ActivityItem) {
-        let projection = source_projection(&item);
-        append_source_projection(&mut item.source_projections, projection);
-        let key = canonical_identity(&item);
-        self.items
-            .entry(key)
-            .and_modify(|existing| merge_item(existing, &item))
-            .or_insert(item);
-    }
-
-    fn items(self, scope: ActivityScope, limit: usize) -> Vec<ActivityItem> {
-        let mut items = self.items.into_values().collect::<Vec<_>>();
-        for item in &mut items {
-            finalize_item(item);
-        }
-        items.sort_by(|left, right| item_sort_key(right).cmp(&item_sort_key(left)));
-        if scope == ActivityScope::ActiveRecent {
-            items.retain(|item| is_active(item.state) || item.finished_at.is_some());
-        }
-        items.truncate(limit.max(1));
-        items
-    }
-}
-
-fn canonical_identity(item: &ActivityItem) -> String {
-    // Lifecycle records own agent-task state. Their durable id is also the
-    // observation run id, so normalize both projections onto that one identity.
-    if item.source_store == "agent-task.lifecycle" {
-        return format!("run:{}", item.id);
-    }
-    if let Some(agent_task_run_id) = &item.refs.agent_task_run_id {
-        return format!("run:{agent_task_run_id}");
-    }
-    if let Some(run_id) = &item.refs.run_id {
-        return format!("run:{run_id}");
-    }
-    if let Some(job_id) = &item.refs.runner_job_id {
-        return format!("runner-job:{job_id}");
-    }
-    format!("item:{}", item.id)
-}
-
-fn merge_item(existing: &mut ActivityItem, incoming: &ActivityItem) {
-    if source_precedence(incoming) > source_precedence(existing)
-        || (source_precedence(incoming) == source_precedence(existing)
-            && incoming.updated_at > existing.updated_at)
-    {
-        let mut replacement = incoming.clone();
-        merge_refs(&mut replacement, existing);
-        append_unique(&mut replacement.artifacts, &existing.artifacts);
-        append_unique(&mut replacement.evidence, &existing.evidence);
-        append_actions(&mut replacement.next_actions, &existing.next_actions);
-        append_source_projections(
-            &mut replacement.source_projections,
-            &existing.source_projections,
-        );
-        *existing = replacement;
-        return;
-    }
-
-    merge_refs(existing, incoming);
-    append_unique(&mut existing.artifacts, &incoming.artifacts);
-    append_unique(&mut existing.evidence, &incoming.evidence);
-    append_actions(&mut existing.next_actions, &incoming.next_actions);
-    append_source_projections(
-        &mut existing.source_projections,
-        &incoming.source_projections,
-    );
-}
-
-fn source_precedence(item: &ActivityItem) -> u8 {
-    source_store_precedence(&item.source_store)
-}
-
-fn source_store_precedence(source_store: &str) -> u8 {
-    match source_store {
-        "agent-task.lifecycle" => 4,
-        "runner.session" => 3,
-        "daemon.jobs-json" => 2,
-        "observation.sqlite" => 1,
-        _ => 0,
-    }
-}
-
-fn finalize_item(item: &mut ActivityItem) {
-    item.artifacts.sort_by(|left, right| {
-        left.uri
-            .cmp(&right.uri)
-            .then_with(|| left.kind.cmp(&right.kind))
-            .then_with(|| left.id.cmp(&right.id))
-    });
-    item.evidence.sort_by(|left, right| {
-        left.uri
-            .cmp(&right.uri)
-            .then_with(|| left.kind.cmp(&right.kind))
-            .then_with(|| left.id.cmp(&right.id))
-    });
-    item.next_actions.sort_by(|left, right| {
-        left.command
-            .cmp(&right.command)
-            .then_with(|| left.label.cmp(&right.label))
-    });
-    item.source_projections.sort_by(|left, right| {
-        source_store_precedence(&right.source_store)
-            .cmp(&source_store_precedence(&left.source_store))
-            .then_with(|| right.updated_at.cmp(&left.updated_at))
-            .then_with(|| left.source_store.cmp(&right.source_store))
-            .then_with(|| left.id.cmp(&right.id))
-    });
-    item.state_conflicts = item
-        .source_projections
-        .iter()
-        .filter(|projection| projection.state != item.state)
-        .map(|projection| ActivityStateConflict {
-            source_store: projection.source_store.clone(),
-            id: projection.id.clone(),
-            state: projection.state,
-        })
-        .collect();
-}
-
-fn merge_refs(existing: &mut ActivityItem, incoming: &ActivityItem) {
-    if existing.refs.run_id.is_none() {
-        existing.refs.run_id = incoming.refs.run_id.clone();
-    }
-    if existing.refs.agent_task_run_id.is_none() {
-        existing.refs.agent_task_run_id = incoming.refs.agent_task_run_id.clone();
-    }
-    if existing.refs.runner_job_id.is_none() {
-        existing.refs.runner_job_id = incoming.refs.runner_job_id.clone();
-    }
-    if existing.runner.runner_id.is_none() {
-        existing.runner.runner_id = incoming.runner.runner_id.clone();
-    }
-    if existing.runner.job_id.is_none() {
-        existing.runner.job_id = incoming.runner.job_id.clone();
-    }
-    if existing.runner.transport.is_none() {
-        existing.runner.transport = incoming.runner.transport.clone();
-    }
-}
-
-fn source_projection(item: &ActivityItem) -> ActivitySourceProjection {
-    ActivitySourceProjection {
-        source_store: item.source_store.clone(),
-        id: item.id.clone(),
-        state: item.state,
-        updated_at: item.updated_at.clone(),
-        finished_at: item.finished_at.clone(),
-    }
-}
-
-fn append_source_projection(
-    target: &mut Vec<ActivitySourceProjection>,
-    incoming: ActivitySourceProjection,
-) {
-    if !target.iter().any(|projection| {
-        projection.source_store == incoming.source_store && projection.id == incoming.id
-    }) {
-        target.push(incoming);
-    }
-}
-
-fn append_source_projections(
-    target: &mut Vec<ActivitySourceProjection>,
-    incoming: &[ActivitySourceProjection],
-) {
-    for projection in incoming {
-        append_source_projection(target, projection.clone());
-    }
-}
-
-fn append_unique(target: &mut Vec<ActivityEvidenceRef>, incoming: &[ActivityEvidenceRef]) {
-    for item in incoming {
-        if !target.iter().any(|existing| existing.uri == item.uri) {
-            target.push(item.clone());
-        }
-    }
-}
-
-fn append_actions(target: &mut Vec<ActivityNextAction>, incoming: &[ActivityNextAction]) {
-    for action in incoming {
-        if !target
-            .iter()
-            .any(|existing| existing.command == action.command)
-        {
-            target.push(action.clone());
-        }
-    }
-}
-
-fn item_sort_key(item: &ActivityItem) -> (bool, Option<DateTime<Utc>>, String) {
-    (
-        is_active(item.state),
-        item.updated_at
-            .as_deref()
-            .or(Some(item.created_at.as_str()))
-            .and_then(parse_ts),
-        item.id.clone(),
-    )
-}
-
-fn parse_ts(value: &str) -> Option<DateTime<Utc>> {
-    DateTime::parse_from_rfc3339(value)
-        .ok()
-        .map(|dt| dt.with_timezone(&Utc))
-}
-
-fn metadata_string(value: &Value, keys: &[&str]) -> Option<String> {
-    keys.iter()
-        .find_map(|key| value.get(*key).and_then(Value::as_str))
-        .filter(|value| !value.trim().is_empty())
-        .map(str::to_string)
-}
-
-fn action(label: impl Into<String>, command: impl Into<String>) -> ActivityNextAction {
-    ActivityNextAction {
-        label: label.into(),
-        command: command.into(),
-    }
-}
-
-fn ms_to_rfc3339(ms: u64) -> String {
-    DateTime::<Utc>::from_timestamp_millis(ms as i64)
-        .unwrap_or_else(Utc::now)
-        .to_rfc3339()
-}
-
-mod observation {
-    use super::*;
-
-    /// Resolve a single observation-run activity item by id without scanning the
-    /// full corpus. Matches the persisted run id directly (indexed `get_run`),
-    /// so `activity show <run-id>` returns immediately for a known run (#9762).
-    pub(super) fn probe_by_id(id: &str) -> Result<Option<ActivityItem>> {
-        let store = ObservationStore::open_initialized()?;
-        match store.get_run(id)? {
-            Some(run) => Ok(Some(item_from_run(&store, run)?)),
-            None => Ok(None),
-        }
-    }
-
-    pub(super) fn collect(collector: &mut ActivityCollector, limit: usize) -> Result<()> {
-        let store = ObservationStore::open_initialized()?;
-        let mut records = store.list_runs(RunListFilter {
-            limit: Some(limit as i64),
-            ..Default::default()
-        })?;
-        let listed_ids = records
-            .iter()
-            .map(|record| record.id.clone())
-            .collect::<BTreeSet<_>>();
-        // Recent terminal records are bounded for display, but active work is
-        // always included before the canonical report applies its final limit.
-        records.extend(
-            store
-                .list_active_runs()?
-                .into_iter()
-                .filter(|record| !listed_ids.contains(&record.id)),
-        );
-        for run in records {
-            collector.insert(item_from_run(&store, run)?);
-        }
-        Ok(())
-    }
-
-    fn item_from_run(store: &ObservationStore, run: RunRecord) -> Result<ActivityItem> {
-        let artifacts = store
-            .list_artifacts(&run.id)?
-            .into_iter()
-            .map(|artifact| ActivityEvidenceRef {
-                id: artifact.id,
-                kind: artifact.kind,
-                uri: artifact
-                    .public_url
-                    .or(artifact.url)
-                    .unwrap_or(artifact.path),
-            })
-            .collect::<Vec<_>>();
-        let runner_id = metadata_string(&run.metadata_json, &["runner_id"]);
-        let job_id = metadata_string(&run.metadata_json, &["runner_job_id", "job_id"]);
-        Ok(ActivityItem {
-            id: run.id.clone(),
-            kind: run.kind.clone(),
-            source_store: "observation.sqlite".to_string(),
-            state: state_from_run_status(&run.status),
-            created_at: run.started_at.clone(),
-            updated_at: run
-                .finished_at
-                .clone()
-                .or_else(|| Some(run.started_at.clone())),
-            finished_at: run.finished_at,
-            command: run.command,
-            cwd: run.cwd,
-            runner: ActivityRunnerRefs {
-                runner_id,
-                job_id: job_id.clone(),
-                transport: None,
-            },
-            refs: ActivityCrossRefs {
-                run_id: Some(run.id.clone()),
-                agent_task_run_id: None,
-                runner_job_id: job_id,
-            },
-            artifacts,
-            evidence: Vec::new(),
-            source_projections: Vec::new(),
-            state_conflicts: Vec::new(),
-            next_actions: actions_for_observation_run(&run.id, state_from_run_status(&run.status)),
-        })
-    }
-
-    pub(super) fn state_from_run_status(status: &str) -> ActivityState {
-        match RunStatus::from_label(status) {
-            Some(RunStatus::Running) => ActivityState::Running,
-            Some(RunStatus::Pass | RunStatus::Skipped) => ActivityState::Succeeded,
-            Some(RunStatus::Fail | RunStatus::Error) => ActivityState::Failed,
-            Some(RunStatus::Stale) => ActivityState::Stale,
-            None => ActivityState::Unknown,
-        }
-    }
-
-    fn actions_for_observation_run(run_id: &str, state: ActivityState) -> Vec<ActivityNextAction> {
-        let mut actions = vec![action("show run", format!("homeboy runs show {run_id}"))];
-        if is_active(state) {
-            actions.push(action(
-                "watch run",
-                format!("homeboy activity watch {run_id}"),
-            ));
-        }
-        actions.push(action(
-            "artifacts",
-            format!("homeboy runs artifacts {run_id}"),
-        ));
-        if matches!(state, ActivityState::Stale) {
-            actions.push(action("reconcile stale runs", "homeboy runs reconcile"));
-        }
-        actions
-    }
-}
-
-mod daemon_jobs {
-    use super::*;
-
-    /// Resolve a single daemon-job activity item by id without listing every
-    /// job. The daemon job id is a UUID; when `id` parses as one, look it up
-    /// directly. Non-UUID ids (run labels, etc.) resolve through other providers
-    /// or the full-scan fallback (#9762).
-    pub(super) fn probe_by_id(id: &str) -> Result<Option<ActivityItem>> {
-        let Ok(job_id) = uuid::Uuid::parse_str(id) else {
-            return Ok(None);
-        };
-        let path = paths::daemon_jobs_file()?;
-        if !path.exists() {
-            return Ok(None);
-        }
-        let store = api_jobs::JobStore::open_without_reconciliation(path)?;
-        match store.get(job_id) {
-            Ok(job) => Ok(Some(item_from_job(&store, job)?)),
-            Err(_) => Ok(None),
-        }
-    }
-
-    pub(super) fn collect(collector: &mut ActivityCollector) -> Result<()> {
-        let path = paths::daemon_jobs_file()?;
-        if !path.exists() {
-            return Ok(());
-        }
-        let store = api_jobs::JobStore::open_without_reconciliation(path)?;
-        for job in store.list() {
-            collector.insert(item_from_job(&store, job)?);
-        }
-        Ok(())
-    }
-
-    pub(super) fn item_from_job(store: &api_jobs::JobStore, job: Job) -> Result<ActivityItem> {
-        let state = ActivityState::from(job.status);
-        let job_id = job.id.to_string();
-        let (durable_run_id, agent_task_run_id) = job_run_refs(store, &job);
-        Ok(ActivityItem {
-            id: job_id.clone(),
-            kind: job.operation.clone(),
-            source_store: "daemon.jobs-json".to_string(),
-            state,
-            created_at: ms_to_rfc3339(job.created_at_ms),
-            updated_at: Some(ms_to_rfc3339(job.updated_at_ms)),
-            finished_at: job.finished_at_ms.map(ms_to_rfc3339),
-            command: Some(job.operation),
-            cwd: None,
-            runner: ActivityRunnerRefs {
-                runner_id: job
-                    .claimed_by_runner_id
-                    .clone()
-                    .or(job.target_runner_id.clone()),
-                job_id: Some(job_id.clone()),
-                transport: Some("daemon".to_string()),
-            },
-            refs: ActivityCrossRefs {
-                run_id: durable_run_id,
-                agent_task_run_id,
-                runner_job_id: Some(job_id.clone()),
-            },
-            artifacts: job
-                .artifacts
-                .into_iter()
-                .map(|artifact| ActivityEvidenceRef {
-                    id: artifact.id,
-                    kind: artifact.mime.unwrap_or_else(|| "artifact".to_string()),
-                    uri: artifact
-                        .url
-                        .or(artifact.path)
-                        .unwrap_or_else(|| "homeboy://artifact/unavailable".to_string()),
-                })
-                .collect(),
-            evidence: Vec::new(),
-            source_projections: Vec::new(),
-            state_conflicts: Vec::new(),
-            next_actions: actions_for_job(None, &job_id, state),
-        })
-    }
-
-    fn job_run_refs(store: &api_jobs::JobStore, job: &Job) -> (Option<String>, Option<String>) {
-        store.events(job.id).unwrap_or_default().iter().fold(
-            (None, None),
-            |(durable, agent_task), event| {
-                (
-                    durable.or_else(|| event_metadata_string(event, &["durable_run_id", "run_id"])),
-                    agent_task.or_else(|| event_metadata_string(event, &["agent_task_run_id"])),
-                )
-            },
-        )
-    }
-
-    fn event_metadata_string(event: &JobEvent, keys: &[&str]) -> Option<String> {
-        keys.iter()
-            .find_map(|key| event.data.as_ref()?.get(*key)?.as_str())
-            .filter(|value| !value.trim().is_empty())
-            .map(str::to_string)
-    }
-
-    pub(super) fn actions_for_job(
-        runner_id: Option<&str>,
-        job_id: &str,
-        state: ActivityState,
-    ) -> Vec<ActivityNextAction> {
-        let mut actions = Vec::new();
-        if let Some(runner_id) = runner_id {
-            actions.push(action(
-                "logs",
-                format!("homeboy runner job logs {runner_id} {job_id}"),
-            ));
-            if is_active(state) {
-                actions.push(action(
-                    "follow logs",
-                    format!("homeboy runner job logs {runner_id} {job_id} --follow"),
-                ));
-            }
-        } else {
-            actions.push(action("daemon status", "homeboy daemon status"));
-        }
-        actions
-    }
-}
-
-mod runner_sessions {
-    use super::*;
-
-    pub(super) fn collect(collector: &mut ActivityCollector) {
-        // Use the latency-bounded indexed view: activity only needs the
-        // current/recent active-job list, never the full generation reconcile
-        // that `statuses()` performs (one blocking poll per draining
-        // generation). This keeps `homeboy activity` bounded as generation
-        // history grows (#9522).
-        for report in crate::observation::runs_service::with_runner_evidence(|provider| {
-            provider.statuses_indexed()
-        })
-        .into_iter()
-        .filter(|report| report.connected)
-        {
-            for job in report.active_jobs {
-                collector.insert(item_from_active_runner_job(job));
-            }
-        }
-    }
-
-    fn item_from_active_runner_job(job: ActiveRunnerJobSummary) -> ActivityItem {
-        let state = if job.stale_reason.is_some() {
-            ActivityState::Stale
-        } else {
-            ActivityState::from(job.status)
-        };
-        ActivityItem {
-            id: job
-                .durable_run_id
-                .clone()
-                .unwrap_or_else(|| job.job_id.clone()),
-            kind: job.kind.clone(),
-            source_store: "runner.session".to_string(),
-            state,
-            created_at: ms_to_rfc3339(job.started_at_ms),
-            updated_at: Some(ms_to_rfc3339(job.updated_at_ms)),
-            finished_at: None,
-            command: Some(job.command.clone()),
-            cwd: job.cwd,
-            runner: ActivityRunnerRefs {
-                runner_id: Some(job.runner_id.clone()),
-                job_id: Some(job.job_id.clone()),
-                transport: Some(job.source.clone()),
-            },
-            refs: ActivityCrossRefs {
-                run_id: job.durable_run_id,
-                agent_task_run_id: None,
-                runner_job_id: Some(job.job_id.clone()),
-            },
-            artifacts: Vec::new(),
-            evidence: Vec::new(),
-            source_projections: Vec::new(),
-            state_conflicts: Vec::new(),
-            next_actions: actions_for_runner_job(&job.runner_id, &job.job_id, state),
-        }
-    }
-
-    fn actions_for_runner_job(
-        runner_id: &str,
-        job_id: &str,
-        state: ActivityState,
-    ) -> Vec<ActivityNextAction> {
-        let mut actions = daemon_jobs::actions_for_job(Some(runner_id), job_id, state);
-        if is_active(state) {
-            actions.push(action("watch", format!("homeboy activity watch {job_id}")));
-        }
-        if is_failure(state) {
-            actions.push(action(
-                "reconcile",
-                format!("homeboy runner job logs {runner_id} {job_id}"),
-            ));
-        }
-        actions
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::api_jobs::JobStatus;
-    use crate::observation::NewRunRecord;
-    use crate::run_lifecycle_record::RunExecutionState;
+    use crate::api_jobs::{self, JobStatus};
+    use crate::observation::{NewRunRecord, ObservationStore, RunStatus};
+    use crate::paths;
     use crate::test_support::with_isolated_home;
 
     fn item(id: &str, state: ActivityState) -> ActivityItem {
